Refactor NBT handling and update buffer types to long
Updated various classes and structs to use 'long' for position and length, enhancing support for larger data sizes. Refactored NBT-related classes to utilize the SharpNBT library for improved tag handling, including implicit conversions and streamlined writing processes. Updated packet classes to efficiently manage NBT data serialization and deserialization. Added SharpNBT library reference to the project file.

+++ src/Api/Mojang/Profiles/IdentifiedKey.cs
@@ -84,7 +84,7 @@ public record IdentifiedKey(IdentifiedKeyRevision Revision, long ExpiresAt, byte
             using var rsa = RSA.Create();
             rsa.ImportSubjectPublicKeyInfo(YggdrasilSessionPublicKey, out _);
 
-            return rsa.VerifyData(verify[..buffer.Position], Signature, HashAlgorithmName.SHA1, RSASignaturePadding.Pkcs1);
+            return rsa.VerifyData(verify[..(int)buffer.Position], Signature, HashAlgorithmName.SHA1, RSASignaturePadding.Pkcs1);
         }
     }
 }
